// src/shared/interfaces/completed-courses.interface.ts
export interface ICompletedCoursesData {
  id: number | null;
  Usuario: number | null;
  Nombre: string | null;
  Apellido: string | null;
  Email: string | null;
  city: string | null;
  country: string | null;
  idCurso: number | null;
  Curso: string | null;
  cursocorto: string | null;
  Categoria: string | null;
  Categoria1: string | null;
  tipo_curso: string | null;
  estado: number | null;
  Curso_Finalizado: number | null;
}